[feature] migrate rsof plotting utils.
